Mac port:  - fixed intro/logo splash loading via resource manager,  - moved resources into .app/Contents/Resource/Data/,  - added a de-appify feature for the build system,  - fixed refresh rate (as seen in the Mac title bar).

+++ UiLepra/Source/Mac/UiMacDisplayManager.mm
@@ -720,7 +720,7 @@ DisplayMode MacDisplayManager::ConvertNativeDisplayMode(CFDictionaryRef pMode)
 	CFNumberGetValue(lResHeight, kCFNumberSInt32Type, &lHeight);
 	int lBpp;
 	CFNumberGetValue(lResBitsPerPixel, kCFNumberSInt32Type, &lBpp);
-	int lRefreshRate;
+	float lRefreshRate;
 	CFNumberGetValue(lResRefreshRate, kCFNumberFloat32Type, &lRefreshRate);
 
 	DisplayMode lDisplayMode;
